# Optimize and simplify Correctness::compute