--- a/src/lib.rs
+++ b/src/lib.rs
@@ -68,24 +68,22 @@ impl Correctness {
         assert_eq!(answer.len(), 5);
         assert_eq!(guess.len(), 5);
         let mut c = [Correctness::Wrong; 5];
-        let mut used = [false; 5];
-        // Mark things green
-        for (i, (a, g)) in answer.bytes().zip(guess.bytes()).enumerate() {
-            if a == g {
+        let answer_bytes = answer.as_bytes();
+        let guess_bytes = guess.as_bytes();
+
+        for (i, &a) in answer_bytes.iter().enumerate() {
+            if a == guess_bytes[i] {
+                // Mark things green
                 c[i] = Correctness::Correct;
-                used[i] = true;
+            } else if let Some(j) = guess_bytes.iter().enumerate().position(|(j, &g)| {
+                // The position in guess can only be marked as Misplaced, if it isn't Correct and wasn't already marked before.
+                a == g && answer_bytes[j] != a && c[j] == Correctness::Wrong
+            }) {
+                // Mark things yellow
+                c[j] = Correctness::Misplaced;
             }
         }
 
-        for (i, g) in guess.bytes().enumerate() {
-            if c[i] == Correctness::Correct {
-                // Already marked as green
-                continue;
-            }
-            if Correctness::is_misplaced(g, answer, &mut used) {
-                c[i] = Correctness::Misplaced;
-            }
-        }
         c
     }