by reducing it's implementation into one loop and removing the used array.
The unit tests helped a lot thanks 😄 
It's mostly a simplification, it improves the performance only a bit.
```
> hyperfine.exe -n old '.\roget-base.exe --implementation cutoff' -n new '.\target\release\roget.exe --implementation cutoff'
Benchmark 1: old
  Time (mean ± σ):      8.366 s ±  0.010 s    [User: 0.000 s, System: 0.006 s]
  Range (min … max):    8.346 s …  8.381 s    10 runs

Benchmark 2: new
  Time (mean ± σ):      7.958 s ±  0.005 s    [User: 0.000 s, System: 0.003 s]
  Range (min … max):    7.949 s …  7.964 s    10 runs

Summary
  'new' ran
    1.05 ± 0.00 times faster than 'old'
```